+++ packages/test/jest.config.ts
@@ -4,7 +4,7 @@ const config: Config.InitialOptions = {
   preset: "ts-jest",
   resetMocks: true,
   testEnvironment: "node",
-  testTimeout: 10000,
+  testTimeout: 20000,
 };
 
 export default config;