+++ packages/test/src/protobuf/opentelemetry/proto/resource/v1/resource.ts
@@ -1,0 +1,60 @@
+/* eslint-disable */
+import _m0 from "protobufjs/minimal";
+import { KeyValue } from "../../common/v1/common";
+
+export const protobufPackage = "opentelemetry.proto.resource.v1";
+
+export interface Resource {
+  attributes: KeyValue[];
+  droppedAttributesCount: number;
+}
+
+function createBaseResource(): Resource {
+  return { attributes: [], droppedAttributesCount: 0 };
+}
+
+export const Resource = {
+  encode(
+    message: Resource,
+    writer: _m0.Writer = _m0.Writer.create(),
+  ): _m0.Writer {
+    for (const v of message.attributes) {
+      KeyValue.encode(v!, writer.uint32(10).fork()).ldelim();
+    }
+    if (message.droppedAttributesCount !== 0) {
+      writer.uint32(16).uint32(message.droppedAttributesCount);
+    }
+    return writer;
+  },
+
+  decode(input: _m0.Reader | Uint8Array, length?: number): Resource {
+    const reader =
+      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseResource();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          if (tag !== 10) {
+            break;
+          }
+
+          message.attributes.push(KeyValue.decode(reader, reader.uint32()));
+          continue;
+        case 2:
+          if (tag !== 16) {
+            break;
+          }
+
+          message.droppedAttributesCount = reader.uint32();
+          continue;
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skipType(tag & 7);
+    }
+    return message;
+  },
+};